chore: restart everything
Removing all of the work I've done so far as the architecture I had wasn't
even faster than the tree-walk implementation. I'm not entirely sure why,
but I think it had something to do with the excessive usage of `last_mut()`
and recursive calls into `run()` and `compile()`. It wasn't any good, but I'm
going to try again.